refactor: 역할에 맞게 메서드 이동
comment: 적절한지 리뷰 필요
